output: Extract candidate preview diffs
Candidate preview output now keeps selector and command formatting outside
the main renderer. The renderer still owns buffer diff printing, including
the same candidate diff text path that already exists in the project.

That diff-printing block depends on core diff parsing, line buffers, and
hunk coloring. Keeping those dependencies in the overview/detail module
makes the renderer look responsible for patch formatting rather than
delegating to it.

This commit moves the existing candidate buffer diff rendering into
output.candidate_preview_diff and updates output.candidate_preview to call
that module.

Subsequent commits will add architecture coverage for the candidate diff
module.
